soundfixes

+++ nes-lib/include/ApuFrameCounter.h
@@ -15,16 +15,16 @@ class ApuFrameCounter {
  private:
   bool mInterruptInhibit = false;
   bool mFiveStepMode = false;
-  unsigned long mFrameCounter = 0;
+  uint32_t mFrameCounter = 0;
   Apu& mApu;
-  const unsigned long mFourStepHalfFrameClocks[2] = {7456, 14914};
-  const unsigned long mFourStepQuarterFrameClocks[4] = {3728, mFourStepHalfFrameClocks[0], 11185,
-                                                        mFourStepHalfFrameClocks[1]};
-  const unsigned long mFiveStepHalfFrameClocks[2] = {7456, 18640};
-  const unsigned long mFiveStepQuarterFrameClocks[4] = {3728, mFiveStepHalfFrameClocks[0], 11185,
-                                                        mFiveStepHalfFrameClocks[1]};
+  const uint32_t mFourStepHalfFrameClocks[2] = {7456, 14914};
+  const uint32_t mFourStepQuarterFrameClocks[4] = {3728, mFourStepHalfFrameClocks[0], 11185,
+                                                   mFourStepHalfFrameClocks[1]};
+  const uint32_t mFiveStepHalfFrameClocks[2] = {7456, 18640};
+  const uint32_t mFiveStepQuarterFrameClocks[4] = {3728, mFiveStepHalfFrameClocks[0], 11185,
+                                                   mFiveStepHalfFrameClocks[1]};
 
-  bool IsHalfFrameClock(int clock);
-  bool IsQuarterFrameClock(int clock);
+  bool IsHalfFrameClock(uint32_t clock);
+  bool IsQuarterFrameClock(uint32_t clock);
   bool HasToSetInterrupt();
 };
